Opening Devices (#157)
* Better rounding and conversion (avoiding casting)

* Bitrates are now longs (very high bitrates result in int overflows). 

* Minor fix in input format detection error message. 

* Display Picture Number and Coded Picture number are now longs. 

* Rounding is now done via System.Convert. 

* It is now possible to open devices! see issue #48

+++ Unosquare.FFME.Common/Decoding/VideoFrame.cs
@@ -106,12 +106,12 @@
         /// If not set by the decoder, this attempts to obtain it by dividing the start time by the
         /// frame duration
         /// </summary>
-        public int DisplayPictureNumber { get; }
+        public long DisplayPictureNumber { get; }
 
         /// <summary>
         /// Gets the coded picture number set by the decoder.
         /// </summary>
-        public int CodedPictureNumber { get; }
+        public long CodedPictureNumber { get; }
 
         /// <summary>
         /// Gets the SMTPE time code.
